[wips] db connection
