+++ Dockerfile.back
@@ -5,4 +5,4 @@ COPY ./requirements.txt /requirements.txt
 RUN pip install -U pip && \
     pip install -r requirements.txt
 
-ENTRYPOINT [ "python", "/app/backend/main.py" ]
+ENTRYPOINT [ "python", "/app/run.py" ]
