--- backend/main.py
@@ -1,10 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-@app.route('/')
-def index():
-  return 'index'
-
-if __name__ == '__main__':
-    app.run(host='0.0.0.0', port=80)
