+++ requirements.txt
@@ -1,4 +1,4 @@
-flask==1.1.1
-autopep8==1.4.4
-flake8==3.7.8
-rope==0.14.0
+flask
+SQLAlchemy
+flask-sqlalchemy
+PyMySQL
